Feature Team as mandatory pick at front end
FIR-307
