+++ src/firefighter/incidents/models/__init__.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from firefighter.incidents.models.component import Component
 from firefighter.incidents.models.environment import Environment
-from firefighter.incidents.models.feature_team import FeatureTeam
 from firefighter.incidents.models.group import Group
 from firefighter.incidents.models.impact import Impact
 from firefighter.incidents.models.incident import Incident
