+++ src/firefighter/raid/models.py
@@ -75,3 +75,28 @@ class RaidArea(models.Model):
 
     def __str__(self) -> str:
         return self.name
+
+
+class FeatureTeam(models.Model):
+    id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=80)
+    jira_project_key = models.CharField(
+        max_length=10,
+        unique=True,
+    )
+
+    class Meta(TypedModelMeta):
+        unique_together = ("name", "jira_project_key")
+        verbose_name = "Feature Team"
+        verbose_name_plural = "Feature Teams"
+
+    def __str__(self) -> str:
+        return self.name
+
+    @property
+    def get_team(self) -> str:
+        return "{self.name}  {self.jira_project_key}"
+
+    @property
+    def get_key(self) -> str:
+        return "{self.jira_project_key}"
